fix: move logics outside of main.rs
